Updates to the nvidia riva runnables to make interrupts possible.
