+++ libs/community/langchain_community/utilities/nvidia_riva/__init__.py
@@ -1,7 +1,14 @@
 """LangChain Expression Language Runnables for performing Riva ASR and TTS in-chain."""
-from .asr import ASRInputType, ASROutputType, RivaASR
+from .asr import (
+    HANGUP,
+    ASRInputType,
+    ASROutputType,
+    AudioStream,
+    RivaASR,
+    StreamInputType,
+    StreamOutputType,
+)
 from .common import RivaAudioEncoding, RivaAuthMixin, RivaCommonConfigMixin, SentinelT
-from .stream import HANGUP, AudioStream, StreamInputType, StreamOutputType
 from .tts import RivaTTS, TTSInputType, TTSOutputType
 
 __all__ = [
